[#898] Small ui improvements
